DAQgrpc/requirements.txt
pip==25.1.1
grpcio
grpcio-tools
pyqt5
pyqtgraph